make safe_overwrite work when target is blank 👻

+++ system_checks.sh
@@ -16,16 +16,18 @@ safe_overwrite() {
   new_file=$(mktemp)
   echo "$2" > "$new_file"
 
-  if ! diff "$1" "$new_file" 1> /dev/null; then
-    moved_to_name="$1.original.$(date +"%Y-%m-%d_%H-%M-%S%Z")"
-    echo -e "\t⚠️  renaming $1 to $moved_to_name"
-    cat << MOVED_NOTE | cat - "$1" > "$moved_to_name"
+  if [ -f "$1" ]; then
+    if ! diff "$1" "$new_file" 1> /dev/null; then
+      moved_to_name="$1.original.$(date +"%Y-%m-%d_%H-%M-%S%Z")"
+      echo -e "\t⚠️  renaming $1 to $moved_to_name"
+      cat << MOVED_NOTE | cat - "$1" > "$moved_to_name"
 ## ~~~~~ ~~~~~ ~~~~~ ~~~~~ ~~~~ THIS FILE WAS MOVED ~~~~ ~~~~~ ~~~~~ ~~~~~ ~~~~~
 ## Original name: ~/.ssh/config
 ## Moved on: $(date +"%Y-%m-%d at %H:%M:%S%Z")
 ## Moved by: $3
 ## ~~~~~ ~~~~~ ~~~~~ ~~~~~ ~~~~ THIS FILE WAS MOVED ~~~~ ~~~~~ ~~~~~ ~~~~~ ~~~~~
 MOVED_NOTE
-    mv "$new_file" "$1"
+    fi
   fi
+  mv "$new_file" "$1"
 }
